fix(ui): keep pure exports off the client barrel and share one sonner instance
