+++ turbo.jsonc
@@ -271,6 +271,10 @@
         // theme.css), so CSS is a real test input; without it a CSS-only change
         // could reuse a cached pass and skip the assertion.
         "src/**/*.css",
+        // Same reasoning for docs: the plugin authoring guide is asserted
+        // against the real tokens, so a docs-only edit must invalidate the
+        // cached result rather than reuse a pass that never checked it.
+        "docs/**/*.md",
         "vitest.config.ts",
         "tsconfig.json"
       ],
